# docs/swagger.yaml
openapi: 3.0.3
info:
  title: TeamSupportPro API
  description: REST API for TeamSupportPro — a team-based IT support ticket management system.
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local development server
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: team_support_session
    apiKeyHeader:
      type: apiKey
      in: header
      name: x-api-key
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Team:
      type: object
      properties:
        id: { type: string, example: "it" }
        name: { type: string, example: "IT Support" }
        code: { type: string, example: "IT" }
        accent: { type: string, example: "#0078d4" }
    Category:
      type: object
      properties:
        id: { type: string, example: "cat-it-hardware" }
        teamId: { type: string, example: "it" }
        name: { type: string, example: "Hardware" }
        description: { type: string, example: "Hardware issues and requests" }
    User:
      type: object
      properties:
        id: { type: string, example: "u-kevin" }
        name: { type: string, example: "Kevin Key" }
        email: { type: string, example: "kevin.key@company.com" }
        teamId: { type: string, example: "it" }
        role: { type: string, enum: [Admin, Staff] }
...
security:
  - cookieAuth: []
  - apiKeyHeader: []
  - bearerAuth: []
paths:
...
  /api/teams:
    get:
      summary: List all teams
      responses:
        "200":
          description: Teams list
          content:
            application/json:
              schema:
                type: object
                properties:
                  teams:
                    type: array
                    items: { $ref: '#/components/schemas/Team' }
    post:
      summary: Create a team (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, code, accent]
              properties:
                id: { type: string }
                name: { type: string }
                code: { type: string }
                accent: { type: string }
      responses:
        "201":
          description: Team created
        "403":
          description: Admin required
  /api/teams/{teamId}:
    get:
      summary: Get a team by ID
      parameters:
        - name: teamId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Team details
          content:
            application/json:
              schema:
                type: object
                properties:
                  team: { $ref: '#/components/schemas/Team' }
        "404":
          description: Team not found
    patch:
      summary: Update a team (Admin only)
      parameters:
        - name: teamId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                code: { type: string }
                accent: { type: string }
      responses:
        "200":
          description: Team updated
        "403":
          description: Admin required
    delete:
      summary: Delete a team (Admin only)
      parameters:
        - name: teamId
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Team deleted
        "403":
          description: Admin required
        "404":
          description: Team not found
  /api/categories:
    get:
      summary: List all categories
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items: { $ref: '#/components/schemas/Category' }
    post:
      summary: Create a category (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [teamId, name]
              properties:
                id: { type: string }
                teamId: { type: string }
                name: { type: string }
                description: { type: string }
      responses:
        "201":
          description: Category created
        "403":
          description: Admin required
  /api/categories/{categoryId}:
    get:
      summary: Get a category by ID
      parameters:
        - name: categoryId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Category details
          content:
            application/json:
              schema:
                type: object
                properties:
                  category: { $ref: '#/components/schemas/Category' }
        "404":
          description: Category not found
    patch:
      summary: Update a category (Admin only)
      parameters:
        - name: categoryId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                teamId: { type: string }
                name: { type: string }
                description: { type: string }
      responses:
        "200":
          description: Category updated
        "403":
          description: Admin required
    delete:
      summary: Delete a category (Admin only)
      parameters:
        - name: categoryId
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Category deleted
        "403":
          description: Admin required
        "404":
          description: Category not found
  /api/users:
    get:
      summary: List all users
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items: { $ref: '#/components/schemas/User' }
    post:
      summary: Create a user (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, teamId, role]
              properties:
                id: { type: string }
                name: { type: string }
                email: { type: string, format: email }
                teamId: { type: string }
                role: { type: string, enum: [Admin, Staff] }
      responses:
        "201":
          description: User created
        "403":
          description: Admin required
  /api/users/{userId}:
    get:
      summary: Get a user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
        "404":
          description: User not found
    patch:
      summary: Update a user (Admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                email: { type: string, format: email }
                teamId: { type: string }
                role: { type: string, enum: [Admin, Staff] }
      responses:
        "200":
          description: User updated
        "403":
          description: Admin required
    delete:
      summary: Delete a user (Admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: User deleted
        "403":
          description: Admin required
        "404":
          description: User not found
...